small changes to the deck creating ui and a bug on the deck creation for int max token. Deck creations needs to be tested

+++ lib/core/services/deck/deck_service.dart
@@ -530,7 +530,7 @@ Future<double> _getDeckMaxTokens (String difficultyTypeid) async
   .eq('difficultyType_id',difficultyTypeid)
   .single();
  
-  return deckTokenReader['max_tokens'];
+  return deckTokenReader['max_tokens'].toDouble();
   }catch (e)
   {
   print('MaxTokens: $e');
